requirements.txt
apex==0.9.10dev
decord==0.6.0
deepspeed==0.12.6
editdistance==0.8.1
einops==0.7.0
evaluate==0.4.1
fairscale==0.4.13
fastapi==0.110.0
flash_attn==2.3.4
ftfy==6.1.3
google_api_python_client==2.119.0
gradio==4.21.0
horovod==0.28.1
huggingface_hub==0.20.3
iopath==0.1.10
matplotlib==3.8.2
omegaconf==2.3.0
openai==1.13.3
opencv_python==4.9.0.80
opendatasets==0.1.22
packaging==24.0
pandas==2.2.1
peft==0.6.2
Pillow==10.1.0
Pillow==10.2.0
pycocoevalcap==1.2
pycocotools==2.0.7
python_magic==0.4.27
PyYAML==6.0.1
PyYAML==6.0.1
ray==2.9.2
regex==2023.12.25
Requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
shortuuid==1.0.12
skimage==0.0
spacy==3.7.4
tensorflow==2.16.1
timm==0.9.12
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.33.0
uvicorn==0.28.0
webdataset==0.2.86
